+++ lib/invoicexpress/client.rb
@@ -18,6 +18,7 @@ require 'invoicexpress/client/debit_notes'
 require 'invoicexpress/client/credit_notes'
 require 'invoicexpress/client/simplified_invoices'
 require 'invoicexpress/client/invoice_receipts'
+require 'invoicexpress/client/transport_guides'
 
 
 module Invoicexpress
@@ -53,6 +54,6 @@ module Invoicexpress
     include Invoicexpress::Client::CreditNotes
     include Invoicexpress::Client::SimplifiedInvoices
     include Invoicexpress::Client::InvoiceReceipts
-    
+    include Invoicexpress::Client::TransportGuides
   end
 end