+++ lib/invoicexpress/request.rb
@@ -22,7 +22,7 @@ module Invoicexpress
     end
 
     private
-    
+
     # Executes the request, checking if ti was successful
     #
     # @return [Boolean] True on success, false otherwise
@@ -47,6 +47,7 @@ module Invoicexpress
 
         case method
         when :get, :delete, :head
+          request.options.params_encoder = Faraday::FlatParamsEncoder 
           request.url(path, options)
         when :patch, :post, :put
           request.headers['Content-Type'] = "application/xml; charset=utf-8"